Added routers and 404 and logger middleware
