+++ Library/middleware/404.js
@@ -1,0 +1,3 @@
+module.exports = (req, res) => {
+    res.status(404).json('The route is unavailable');
+}
