+++ Library/routers/books.js
@@ -1,0 +1,122 @@
+const express = require('express');
+const { v4: uuid } = require('uuid');
+
+class Book {
+    constructor(
+        title = "",
+        description = "",
+        authors = "",
+        favorite = "",
+        fileCover = "",
+        fileName = "",
+        id = uuid()
+    ) {
+        this.id = id;
+        this.title = title;
+        this.description = description;
+        this.authors = authors;
+        this.favorite = favorite;
+        this.fileCover = fileCover;
+        this.fileName = fileName;
+    };
+}
+
+const storage = {
+    books: []
+}
+
+const router = express.Router();
+
+router.get('/', (req, res) => {
+    res.json(storage.books);
+});
+
+
+router.get('/:id', (req, res) => {
+    const { id } = req.params;
+    const { books } = storage;
+
+    for (let book of books) {
+        if (book.id === id) {
+            res.json(book);
+            return;
+        }
+    }
+
+    res.status(404).send('The book is not found');
+});
+
+
+router.post('/', (req, res) => {
+    const { books } = storage;
+    const {
+        title,
+        description,
+        authors,
+        favorite,
+        fileCover,
+        fileName
+    } = req.body;
+
+    const book = new Book(
+        title,
+        description,
+        authors,
+        favorite,
+        fileCover,
+        fileName
+    );
+
+    books.push(book);
+    res.status(201).json(books.at(-1));
+});
+
+
+router.put('/:id', (req, res) => {
+    const { id } = req.params;
+    const { books } = storage;
+    const {
+        title,
+        description,
+        authors,
+        favorite,
+        fileCover,
+        fileName
+    } = req.body;
+
+    const idx = books.findIndex(el => el.id === id)
+
+    if (idx !== -1) {
+        books[idx] = {
+            ...books[idx],
+            title,
+            description,
+            authors,
+            favorite,
+            fileCover,
+            fileName
+        };
+
+
+        res.json(books[idx]);
+    } else {
+        res.status(404).send('The book is not found');
+    }
+});
+
+
+router.delete('/:id', (req, res) => {
+    const { id } = req.params;
+    const { books } = storage;
+
+    const idx = books.findIndex(el => el.id === id);
+
+    if (idx !== -1) {
+        books.splice(idx, 1);
+    }
+
+    res.send('ok');
+});
+
+
+module.exports = router;
